requirements.txt
# Django核心
Django==5.0.6
pillow==10.0.0

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
drf-spectacular==0.27.0

# 媒体处理
moviepy==1.0.3

# HTTP请求
requests==2.31.0
urllib3==2.1.0

# 数据库（可选PostgreSQL）
# psycopg2-binary==2.9.9

# 云存储（可选）
# django-storages==1.14.2
# boto3==1.34.0  # AWS S3
# oss2==2.18.0   # 阿里云OSS

# 部署相关
uwsgi==2.0.23
python-dotenv==1.0.0

# 开发工具
ipython==8.12.0

